+++ src/horizon/horizon_api.ts
@@ -1,4 +1,4 @@
-import { AssetType, MemoType } from "stellar-base";
+import { AssetType, MemoType } from "@stellar/stellar-base";
 
 /* tslint:disable-next-line:no-namespace */
 export namespace HorizonApi {