+++ src/horizon/payment_call_builder.ts
@@ -5,13 +5,20 @@ import { ServerApi } from "./server_api";
  * Creates a new {@link PaymentCallBuilder} pointed to server defined by serverUrl.
  *
  * Do not create this object directly, use {@link Server#payments}.
- * @see [All Payments](https://developers.stellar.org/api/resources/payments/)
+ * @see [All Payments](https://developers.stellar.org/api/horizon/resources/list-all-payments/)
  * @constructor
  * @extends CallBuilder
  * @param {string} serverUrl Horizon server URL.
  */
 export class PaymentCallBuilder extends CallBuilder<
-  ServerApi.CollectionPage<ServerApi.PaymentOperationRecord>
+  ServerApi.CollectionPage<
+    | ServerApi.PaymentOperationRecord
+    | ServerApi.CreateAccountOperationRecord
+    | ServerApi.AccountMergeOperationRecord
+    | ServerApi.PathPaymentOperationRecord
+    | ServerApi.PathPaymentStrictSendOperationRecord
+    | ServerApi.InvokeHostFunctionOperationRecord
+  >
 > {
   constructor(serverUrl: URI) {
     super(serverUrl, "payments");
@@ -20,7 +27,7 @@ export class PaymentCallBuilder extends CallBuilder<
 
   /**
    * This endpoint responds with a collection of Payment operations where the given account was either the sender or receiver.
-   * @see [Payments for Account](https://developers.stellar.org/api/resources/accounts/payments/)
+   * @see [Payments for Account](https://developers.stellar.org/api/horizon/resources/get-payments-by-account-id)
    * @param {string} accountId For example: `GDGQVOKHW4VEJRU2TETD6DBRKEO5ERCNF353LW5WBFW3JJWQ2BRQ6KDD`
    * @returns {PaymentCallBuilder} this PaymentCallBuilder instance
    */
@@ -30,7 +37,7 @@ export class PaymentCallBuilder extends CallBuilder<
 
   /**
    * This endpoint represents all payment operations that are part of a valid transactions in a given ledger.
-   * @see [Payments for Ledger](https://developers.stellar.org/api/resources/ledgers/payments/)
+   * @see [Payments for Ledger](https://developers.stellar.org/api/horizon/resources/retrieve-a-ledgers-payments)
    * @param {number|string} sequence Ledger sequence
    * @returns {PaymentCallBuilder} this PaymentCallBuilder instance
    */