+++ src/horizon/types/assets.ts
@@ -1,4 +1,4 @@
-import { AssetType } from "stellar-base";
+import { AssetType } from "@stellar/stellar-base";
 import { HorizonApi } from "./../horizon_api";
 
 export interface AssetRecord extends HorizonApi.BaseResponse {