+++ src/soroban/parsers.ts
@@ -1,4 +1,4 @@
-import { xdr, Contract, SorobanDataBuilder } from 'stellar-base';
+import { xdr, Contract, SorobanDataBuilder } from '@stellar/stellar-base';
 import { Api } from './api';
 
 export function parseRawSendTransaction(
@@ -23,11 +23,15 @@ export function parseRawEvents(
   return {
     latestLedger: r.latestLedger,
     events: (r.events ?? []).map((evt) => {
+      const clone: Omit<Api.RawEventResponse, 'contractId'> = { ...evt };
+      delete (clone as any).contractId; // `as any` hack because contractId field isn't optional
+
+      // the contractId may be empty so we omit the field in that case
       return {
-        ...evt,
-        contractId: new Contract(evt.contractId),
+        ...clone,
+        ...(evt.contractId !== '' && { contractId: new Contract(evt.contractId) }),
         topic: evt.topic.map((topic) => xdr.ScVal.fromXDR(topic, 'base64')),
-        value: xdr.ScVal.fromXDR(evt.value.xdr, 'base64')
+        value: xdr.ScVal.fromXDR(evt.value, 'base64')
       };
     })
   };
@@ -49,7 +53,9 @@ export function parseRawLedgerEntries(
         lastModifiedLedgerSeq: rawEntry.lastModifiedLedgerSeq,
         key: xdr.LedgerKey.fromXDR(rawEntry.key, 'base64'),
         val: xdr.LedgerEntryData.fromXDR(rawEntry.xdr, 'base64'),
-        expirationLedgerSeq: rawEntry.expirationLedgerSeq
+        ...(rawEntry.liveUntilLedgerSeq !== undefined && {
+          liveUntilLedgerSeq: rawEntry.liveUntilLedgerSeq
+        })
       };
     })
   };