+++ test/unit/server/soroban/get_contract_data_test.js
@@ -1,4 +1,4 @@
-const { xdr, nativeToScVal, hash } = StellarSdk;
+const { Address, xdr, nativeToScVal, hash } = StellarSdk;
 const { Server, AxiosClient, Durability } = StellarSdk.SorobanRpc;
 
 describe("Server#getContractData", function () {
@@ -18,7 +18,7 @@ describe("Server#getContractData", function () {
   const ledgerEntry = xdr.LedgerEntryData.contractData(
     new xdr.ContractDataEntry({
       ext: new xdr.ExtensionPoint(0),
-      contract: new StellarSdk.Address(address).toScAddress(),
+      contract: new Address(address).toScAddress(),
       durability: xdr.ContractDataDurability.persistent(),
       key,
       val: key, // lazy
@@ -34,14 +34,10 @@ describe("Server#getContractData", function () {
     }),
   );
 
-  const ledgerExpirationKey = xdr.LedgerKey.expiration(
-    new xdr.LedgerKeyExpiration({ keyHash: hash(ledgerKey.toXDR()) }),
-  );
-
-  const ledgerExpirationEntry = xdr.LedgerEntryData.expiration(
-    new xdr.ExpirationEntry({
+  const ledgerTtlEntry = xdr.LedgerEntryData.ttl(
+    new xdr.TtlEntry({
       keyHash: hash(ledgerKey.toXDR()),
-      expirationLedgerSeq: 1000,
+      liveUntilLedgerSeq: 1000,
     }),
   );
 
@@ -50,7 +46,7 @@ describe("Server#getContractData", function () {
       lastModifiedLedgerSeq: 1,
       key: ledgerKey,
       val: ledgerEntry,
-      expirationLedgerSeq: 1000,
+      liveUntilLedgerSeq: 1000,
     };
 
     this.axiosMock
@@ -59,9 +55,7 @@ describe("Server#getContractData", function () {
         jsonrpc: "2.0",
         id: 1,
         method: "getLedgerEntries",
-        params: [
-          [ledgerKey.toXDR("base64"), ledgerExpirationKey.toXDR("base64")],
-        ],
+        params: [[ledgerKey.toXDR("base64")]],
       })
       .returns(
         Promise.resolve({
@@ -70,15 +64,11 @@ describe("Server#getContractData", function () {
               latestLedger: 420,
               entries: [
                 {
+                  liveUntilLedgerSeq: ledgerTtlEntry.ttl().liveUntilLedgerSeq(),
                   lastModifiedLedgerSeq: result.lastModifiedLedgerSeq,
                   key: ledgerKey.toXDR("base64"),
                   xdr: ledgerEntry.toXDR("base64"),
                 },
-                {
-                  lastModifiedLedgerSeq: result.lastModifiedLedgerSeq,
-                  key: ledgerExpirationKey.toXDR("base64"),
-                  xdr: ledgerExpirationEntry.toXDR("base64"),
-                },
               ],
             },
           },
@@ -90,52 +80,7 @@ describe("Server#getContractData", function () {
       .then(function (response) {
         expect(response.key.toXDR("base64")).to.eql(result.key.toXDR("base64"));
         expect(response.val.toXDR("base64")).to.eql(result.val.toXDR("base64"));
-        expect(response.expirationLedgerSeq).to.eql(1000);
-        done();
-      })
-      .catch((err) => done(err));
-  });
-
-  it("expiration entry not present for contract data key in server response", function (done) {
-    let result = {
-      lastModifiedLedgerSeq: 1,
-      key: ledgerKey,
-      val: ledgerEntry,
-    };
-
-    this.axiosMock
-      .expects("post")
-      .withArgs(serverUrl, {
-        jsonrpc: "2.0",
-        id: 1,
-        method: "getLedgerEntries",
-        params: [
-          [ledgerKey.toXDR("base64"), ledgerExpirationKey.toXDR("base64")],
-        ],
-      })
-      .returns(
-        Promise.resolve({
-          data: {
-            result: {
-              latestLedger: 420,
-              entries: [
-                {
-                  lastModifiedLedgerSeq: result.lastModifiedLedgerSeq,
-                  key: result.key.toXDR("base64"),
-                  xdr: result.val.toXDR("base64"),
-                },
-              ],
-            },
-          },
-        }),
-      );
-
-    this.server
-      .getContractData(address, key, Durability.Persistent)
-      .then(function (response) {
-        expect(response.key.toXDR("base64")).to.eql(result.key.toXDR("base64"));
-        expect(response.val.toXDR("base64")).to.eql(result.val.toXDR("base64"));
-        expect(response.expirationLedgerSeq).to.be.undefined;
+        expect(response.liveUntilLedgerSeq).to.eql(1000);
         done();
       })
       .catch((err) => done(err));
@@ -148,22 +93,13 @@ describe("Server#getContractData", function () {
       .contractData()
       .durability(xdr.ContractDataDurability.temporary());
 
-    const ledgerExpirationKeyDupe = xdr.LedgerKey.expiration(
-      new xdr.LedgerKeyExpiration({ keyHash: hash(ledgerKeyDupe.toXDR()) }),
-    );
-
     this.axiosMock
       .expects("post")
       .withArgs(serverUrl, {
         jsonrpc: "2.0",
         id: 1,
         method: "getLedgerEntries",
-        params: [
-          [
-            ledgerKeyDupe.toXDR("base64"),
-            ledgerExpirationKeyDupe.toXDR("base64"),
-          ],
-        ],
+        params: [[ledgerKeyDupe.toXDR("base64")]],
       })
       .returns(Promise.resolve({ data: { result: { entries: [] } } }));