+++ test/unit/server/soroban/request_airdrop_test.js
@@ -116,8 +116,8 @@ describe("Server#requestAirdrop", function () {
       }),
     );
 
-    const ledgerExpirationKey = xdr.LedgerKey.expiration(
-      new xdr.LedgerKeyExpiration({ keyHash: hash(accountKey.toXDR()) }),
+    const ledgerTtlKey = xdr.LedgerKey.ttl(
+      new xdr.LedgerKeyTtl({ keyHash: hash(accountKey.toXDR()) }),
     );
 
     this.axiosMock
@@ -126,9 +126,7 @@ describe("Server#requestAirdrop", function () {
         jsonrpc: "2.0",
         id: 1,
         method: "getLedgerEntries",
-        params: [
-          [accountKey.toXDR("base64"), ledgerExpirationKey.toXDR("base64")],
-        ],
+        params: [[accountKey.toXDR("base64")]],
       })
       .returns(
         Promise.resolve({